backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
pytesseract
Pillow
openai
python-multipart
python-dotenv
requests
sqlalchemy
aiosqlite
